feature: get all order

+++ module/feature/product/domain/response.go
@@ -8,7 +8,7 @@ import (
 type ProductsResponse struct {
 	ID           uint64                 `json:"id"`
 	Name         string                 `json:"name"`
-	Price        float64                `json:"price"`
+	Price        uint64                 `json:"price"`
 	Description  string                 `json:"description"`
 	Photos       []ProductPhotoResponse `json:"photos"`
 	Rating       float64                `json:"rating"`
